# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/movie_apps": {
            "get": {
                "tags": [
                    "MovieApps"
                ],
                "summary": "Get a listing of the MovieApps.",
                "description": "Get all MovieApps",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/MovieApp"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "MovieApps"
                ],
                "summary": "Store a newly created MovieApps in storage",
                "description": "Store MovieApps",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "MovieApps that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/MovieApp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/MovieApp"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/movie_apps/{id}": {
            "get": {
                "tags": [
                    "MovieApps"
                ],
                "summary": "Display the specified MovieApps",
                "description": "Get MovieApps",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of MovieApps",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/MovieApp"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "MovieApps"
                ],
                "summary": "Update the specified MovieApp in storage",
                "description": "Update MovieApp",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of MovieApp",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "MovieApp that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/MovieApp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/MovieApp"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "MovieApps"
                ],
                "summary": "Remove the specified MovieApp from storage",
                "description": "Delete MovieApp",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of MovieApp",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "MovieApp": {
            "required": [
                "name",
                "totalCnimaHall"
            ],
            "properties": {
                "title": {
                    "description": "title",
                    "type": "string"
                },
                "description": {
                    "description": "description",
                    "type": "string"
                },
                "duration": {
                    "description": "duration",
                    "type": "string",
                    "format": "date-time"
                },
                "releaseDate": {
                    "description": "releaseDate",
                    "type": "string",
                    "format": "date-time"
                },
                "country": {
                    "description": "country",
                    "type": "string"
                },
                "genre": {
                    "description": "genre",
                    "type": "string"
                }
            }
        },
...
    }
}